+++ components/layout/NavDesktop.module.css
@@ -42,10 +42,6 @@
   align-items: center;
 }
 
-.navSocials {
-  font-size: 2.5em;
-}
-
 .navLink {
   cursor: pointer;
   font-size: 1em;
