+++ components/welcome/Welcome.module.css
@@ -1,0 +1,7 @@
+.socials {
+  font-size: 4em;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+}
