+++ components/welcome/Welcome.tsx
@@ -1,0 +1,14 @@
+import { AboutInfo } from "../about/AboutInfo"
+import { LayoutColumn } from "../layout/LayoutColumn"
+import Hero from "./Hero"
+
+export const Welcome = () => {
+  return (
+    <>
+      <Hero />
+      <LayoutColumn>
+        <AboutInfo />
+      </LayoutColumn>
+    </>
+  )
+}
