+++ pages/index.tsx
@@ -1,24 +1,9 @@
-import Hero from '../components/Hero'
-import { SocialButton } from '../components/SocialButton'
 import Layout from '../components/layout/Layout'
-import { LayoutColumn } from '../components/layout/LayoutColumn'
-import styles from '../components/layout/NavDesktop.module.css'
-import { Social } from '../data/links'
+import { Welcome } from '../components/welcome/Welcome'
 
 const IndexPage = () => (
   <Layout column={false}>
-    <Hero />
-    <LayoutColumn>
-      <section className={styles.navSocials}>
-        {Social.map(link =>
-          <SocialButton
-            key={link.url}
-            img={link.imageSrc}
-            url={link.url}
-          />
-        )}
-      </section>
-    </LayoutColumn>
+    <Welcome />
   </Layout>
 )
